+++ src/AngleSharp/Html/Dom/Internal/HtmlNoScriptElement.cs
@@ -8,14 +8,14 @@ namespace AngleSharp.Html.Dom
     /// </summary>
     sealed class HtmlNoScriptElement : HtmlElement
     {
-        public HtmlNoScriptElement(Document owner, String? prefix = null)
-            : base(owner, TagNames.NoScript, prefix, GetFlags(owner.Context))
+        public HtmlNoScriptElement(Document owner, String? prefix = null, Boolean? scripting = false)
+            : base(owner, TagNames.NoScript, prefix, GetFlags(scripting ?? owner.Context.IsScripting()))
         {
         }
 
-        private static NodeFlags GetFlags(IBrowsingContext context)
+        private static NodeFlags GetFlags(Boolean scripting)
         {
-            if (context.IsScripting())
+            if (scripting)
             {
                 return NodeFlags.Special | NodeFlags.LiteralText;
             }